Rename "test" dir to "tests"
