+++ Makefile
@@ -23,11 +23,11 @@ all: venv images docker-compose.yml test
 # Run the tests with testinfra (actually our custom wrapper at ./bin/testinfra)
 # REF: http://testinfra.readthedocs.io/en/latest/
 test: lint all
-	testinfra -v test/
+	testinfra -v tests/
 .PHONY: test
 
 lint: venv
-	flake8 test/
+	flake8 tests/
 
 docker-compose.yml: venv
 	jinja2 \
